+++ src/main/java/org/edu/infi_payment_system/Transaction/service/TransactionLedgerService.java
@@ -1,11 +1,12 @@
 package org.edu.infi_payment_system.Transaction.service;
 
-import org.edu.infi_payment_system.Transaction.dto.TransactionRequestDto;
 import org.edu.infi_payment_system.Transaction.dto.TransactionResponseDto;
+
+import java.math.BigDecimal;
 
 public interface TransactionLedgerService {
     TransactionResponseDto createDoubleEntryTransaction(Long transactionId,
                                                         Long senderId,
                                                         Long receiverId,
-                                                        Double amount);
+                                                        BigDecimal amount);
 }
